# docs/swagger.yaml
basePath: /
definitions:
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Reusable Go REST API backend template.
  title: WarSpark API
  version: dev
paths:
...
  /api/v1/builder-base-leagues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取夜世界联赛列表
      tags:
      - capital
  /api/v1/builder-base-leagues/{id}:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取夜世界联赛详情
      tags:
      - capital
  /api/v1/capital-leagues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取都城联赛列表
      tags:
      - capital
  /api/v1/capital-leagues/{id}:
    get:
      parameters:
      - description: 联赛ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取都城联赛详情
      tags:
      - capital
...
  /api/v1/clans/{tag}:
    get:
      parameters:
      - description: 部落标签
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落信息
      tags:
      - clan
  /api/v1/clans/{tag}/capital-raid-seasons:
    get:
      parameters:
      - description: 部落标签
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取都城突袭赛季
      tags:
      - capital
...
  /api/v1/clans/labels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落标签
      tags:
      - label
...
  /api/v1/players/labels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家标签
      tags:
      - label
...
schemes:
- http
swagger: "2.0"